backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
presidio-analyzer>=2.2.354
presidio-anonymizer>=2.2.354
spacy>=3.7,<3.8
hl7apy==1.3.4
pdfplumber==0.10.3
pytesseract==0.3.13
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
python-multipart==0.0.6
Pillow>=10.0
aiofiles==23.2.1
